Refactor shopping list management with improved sorting and export/import
